back-end/requirements.txt
fastapi[standard]
pillow
timm
torch
transformers
uvicorn
